+++ src/types/bibata-live.d.ts
@@ -10,7 +10,13 @@ declare module 'bibata/app' {
   type Color = {
     base: string;
     outline: string;
-    watch?: string;
+    watch?: {
+      bg?: string;
+      c1?: string;
+      c2?: string;
+      c3?: string;
+      c4?: string;
+    };
   };
 
   type Colors = {