+++ src/components/Cursors/card.tsx
@@ -158,7 +158,7 @@ Report Issue here: https://github.com/ful1e5/bibata/issues`
       }}>
       <div
         style={{
-          backgroundColor: `color-mix(in srgb, #1e1e1e 95%, ${props.color.base})`
+          backgroundColor: `color-mix(in srgb, #1f1f1f 97%, ${props.color.base})`
         }}
         className='w-full mb-4 overflow-hidden rounded-2xl sm:rounded-3xl border-white/[.1] border'>
         <div title={name} className='relative w-full h-24 sm:h-40'>