+++ src/components/Cursors/error.tsx
@@ -1,6 +1,7 @@
 'use client';
 
 import React, { useState } from 'react';
+
 import Link from 'next/link';
 
 import { bugReportTemplate } from '@utils/bug-report';