+++ src/components/DownloadButton/index.tsx
@@ -9,7 +9,7 @@ import { DownloadCount } from './counts';
 import { DownloadSponsor } from './sponsor';
 import { DownloadSubButtons } from './sub-buttons';
 import { DownloadError } from './error';
-import { LockSVG, ProcessingSVG } from '@components/svgs';
+import { DownloadSVG, LockSVG, ProcessingSVG } from '@components/svgs';
 
 import { Platform, Type } from '@prisma/client';
 import { AddDownloadData } from '@services/download';
@@ -232,14 +232,20 @@ export const DownloadButton: React.FC<Props> = (props) => {
               ? 'Download locked while collecting cursor images.'
               : 'Download'
           }
-          className='disabled:opacity-50 relative flex justify-center items-center gap-2 w-4/5 sm:w-1/3 lg:w-1/5 h-11 sm:h-16 rounded-2xl sm:rounded-3xl py-3 bg-green-600 hover:bg-green-500'
+          className='relative flex justify-center items-center uppercase gap-2 w-4/5 sm:w-1/3 lg:w-1/5 h-16 sm:h-20 rounded-full bg-green-600 hover:bg-green-500'
           disabled={props.disabled && !lock && !props.lock}
           onClick={() => !props.lock && setShowDropdown(!showDropdown)}>
-          <p className='overflow-auto text-sm sm:text-lg font-semibold'>
+          {props.lock ? (
+            <LockSVG />
+          ) : busy ? (
+            <ProcessingSVG />
+          ) : (
+            <DownloadSVG />
+          )}
+
+          <p className='overflow-auto text-md font-bold'>
             {!props.lock && busy ? 'Processing' : 'Download'}
           </p>
-
-          {props.lock ? <LockSVG /> : busy && <ProcessingSVG />}
         </button>
       </div>