+++ src/components/NavBar/index.tsx
@@ -7,7 +7,7 @@ import { usePathname } from 'next/navigation';
 import { useSession } from 'next-auth/react';
 
 import { Profile } from './profile';
-import { ProBadge, BibataLogo } from '@components/svgs';
+import { BibataTypoLogo } from '@components/svgs';
 
 type Props = {};
 
@@ -17,16 +17,24 @@ export const NavBar: React.FC<Props> = (_props) => {
   const { data: session, status } = useSession();
 
   return (
-    <header className='sticky py-px top-0 z-20 bg-[#030303] backdrop-filter backdrop-blur-2xl border-b border-white/[.1] firefox:bg-opacity-70 bg-opacity-70'>
+    <header
+      className={`sticky py-px top-0 z-20 ${
+        pathname === '/'
+          ? ''
+          : 'bg-[#151515] backdrop-filter backdrop-blur-xl border-b border-white/[.1] firefox:bg-opacity-95 bg-opacity-95'
+      } `}>
       <nav className='container mx-auto p-3 md:p-4 flex items-center justify-between'>
         <Link href='/'>
           <div className='overflow-hidden flex items-center justify-center gap-2'>
-            <BibataLogo />
             <span className='inline-flex items-center gap-1'>
-              {pathname === '/studio' && (
-                <span className='text-2xl font-bold'>Studio</span>
-              )}
-              {session?.user?.role === 'PRO' && <ProBadge size={18} />}
+              <span
+                className={
+                  session?.user?.role === 'PRO' || pathname === '/'
+                    ? 'text-[#d7f47e]'
+                    : 'text-white'
+                }>
+                <BibataTypoLogo />
+              </span>
             </span>
           </div>
         </Link>