+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### What's New?
 
--   Monochrome colored wedge animation Mode in Customize button
+-   feature: Monochrome colored wedge animation Mode in Customize button
+-   ui: Tint background color on `studio` page
 
 ## [v1.0.0-alpha.1] - 07 December 2023